Moved copy_to to impl and move template plugin to this API
