+++ src/core/include/openvino/runtime/tensor.hpp
@@ -168,7 +168,7 @@ public:
      *
      * @param dst destination tensor
      */
-    void copy_to(ov::Tensor& dst) const;
+    void copy_to(ov::Tensor dst) const;
 
     /**
      * @brief Reports whether the tensor is continuous or not
